Added directory coverage values for better treemap support
The SonarQube treemap widgets work better if metrics exist for the
project, the directory and the file level.

Project coverage is defined as sum of ALL children
Directory coverage is defined as sum of the DIRECT children

The directories have to be treated special because the SonarQube treemap
displays all directories on the same level (not as a tree). Using full
sums would hide bad coverage results in some cases.