Search rule by key
For custom rules created from a template rule, the `internalKey` is set
to the internal key of the template rule by Sonar and not the custom
rule key. When looking up the rule by internal key and providing the
custom rule key, nothing is found as they do not match.